added a part of the allocation algorithm

+++ src/main/java/com/app/cfp/dto/MedicalCaseFullDTO.java
@@ -31,7 +31,7 @@ public class MedicalCaseFullDTO {
 
     private Date insertDate;
 
-    private float score;
+    private double score;
 
     private ResidentDTO resident;
